lib-gralloc-qsd8k: Add support for async framebuffer display updates
Added a thread to run synchronously with FB and MDP driver for framebuffer
display updates. Main thread runs synchronously with upstream callers
of fb_post() such as surfaceflinger.  This allows swapInterval 0 to be easily
supported in fb_post().  It also  decouples compostion from display allowing
better performance especially in MDDI mode and with swapInterval set to 0.