Don't force notifications out the speaker.
APM forces notification sounds out the speaker, even when you have
headphones plugged in.  This is bad; if I'm on headphones, that means
I don't want noise out the speaker to bother others--and worse, I
may not even be aware of it (since it plays out the headphones too).

This also fixes a bigger problem: notifications when on headphones cause
the music to be muted.  Other code in APM does this to prevent music
from playing out the speaker when it's turned on for the notification.

This only fixes this for MSM7K hardware (G1).  The N1 APM code is
somewhat more iterated, with more logic for BlueTooth; the same change
could be made but I can't test how it'd affect BT use cases.  (Actually,
a lot of that new logic may want to be backported to the MSM7K code;
once I get some BT hardware to test with I'll find out.)