Don't force notifications out the speaker (N1).
I take that back: the N1 logic is just like the G1's, except it happens
to already be disabled when on a car dock.  I'm pretty sure this change
is fine, but it still needs to be tested on N1.

(replicates 3303cdc0bc0eb3a0b91188819aabfcdf164c1107 for N1)