librpc: Use infinite timeout for select call.
In clnt.c, select uses timeout value of 500ms.
This way, if any client added/deleted, will be
included/excluded in the next iteration after
the timeout.  But this polling resulsts in more
power consumption.  Use infinite timeout for select
call and asynchronously wake it up using a pipe when
a client is added/deleted.

Change-Id: I80b80ac0c159dda49125fa47fd45f2d039103b86